+++ packages/component/src/hooks/internal/useDictateContext.js
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+
+import Context from '../../Context';
+
+export default function useDictateContext() {
+  return useContext(Context);
+}