+++ packages/component/src/hooks/useAbortable.js
@@ -1,0 +1,7 @@
+import useDictateContext from './internal/useDictateContext';
+
+export default function useAbortable() {
+  const { abortable } = useDictateContext();
+
+  return [abortable];
+}