+++ packages/component/src/hooks/useSupported.js
@@ -1,0 +1,7 @@
+import useDictateContext from './internal/useDictateContext';
+
+export default function useSupported() {
+  const { supported } = useDictateContext();
+
+  return [supported];
+}