+++ packages/component/src/index.js
@@ -1,7 +1,11 @@
 import Composer from './Composer';
+import Context from './Context';
 import DictateButton from './DictateButton';
 import DictateCheckbox from './DictateCheckbox';
+import useAbortable from './hooks/useAbortable';
+import useReadyState from './hooks/useReadyState';
+import useSupported from './hooks/useSupported';
 
-export default DictateButton
+export default DictateButton;
 
-export { Composer, DictateCheckbox }
+export { Composer, Context, DictateCheckbox, useAbortable, useReadyState, useSupported };